+++ CHANGELOG.md
@@ -55,6 +55,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GEOD-1036](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1036)
 - Adjusted Mobile Settings for the List vs Map Tabs
   [GEOD-1037](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1037)
+- Move Popup slightly when opening, whether from map click or accordion selection
+  [GEOD-1219](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1219)
 
 ### Fixed
 - Exploratory Analysis API GET Commands and Interface Behavior